refactor: seperate editors to layout2d/layout3d/decoration.
